Refactor stuff into kerberos.rs
